More warning fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@23351 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/plus4/plus4mem.c
@@ -970,7 +970,7 @@ void mem_inject(DWORD addr, BYTE value)
     /* just call mem_store() to be safe.
        This could possibly be changed to write straight into the
        memory array.  mem_ram[addr & mask] = value; */
-    mem_store(addr & 0xffff, value);
+    mem_store((WORD)(addr & 0xffff), value);
 }
 
 /* ------------------------------------------------------------------------- */
